issue 4 v3
явно указал ErrorParser
