+++ src/main/java/ru/naumen/sd40/log/parser/SdngParser.java
@@ -1,0 +1,14 @@
+package ru.naumen.sd40.log.parser;
+
+import java.io.IOException;
+import java.text.ParseException;
+
+public class SdngParser implements DataParser {
+	
+	public void parseData(DataSet data, String line) throws  IOException, ParseException
+    {
+    	data.parseLineError(line);
+    	data.parseLineAction(line);
+    }
+
+}
