--- src/main/java/ru/naumen/sd40/log/parser/TimeParserInterface.java
@@ -1,9 +1,0 @@
-package ru.naumen.sd40.log.parser;
-
-import java.text.ParseException;
-
-public interface TimeParserInterface {
-	
-	public long parseTime(String line) throws ParseException;
-
-}
